refactor: add quiz controller feat: edit quiz will delete all takequiz record and mail to user
