+++ Views/Quiz/EditQuiz.cshtml
@@ -13,7 +13,7 @@
         <i class="fa-solid fa-circle-plus fs-5"></i>
     </div>
 
-    <form asp-action="EditQuiz">
+    <form asp-controller="Quiz" asp-action="EditQuiz">
         <input type="hidden" asp-for="Id" />
 
         <div class="quiz-name">
